refactor stats related
