+++ libs/common/src/data-analysis/index.ts
@@ -3,4 +3,6 @@ export * from './datamart.service';
 export * from './match-cube-item.interface';
 export * from './match-service-helper';
 export * from './metrics';
+export * from './stats.service';
+export * from './stats.util';
 export * from './time-dimension';
