+++ libs/common/src/services/index.ts
@@ -1,3 +1,2 @@
 export * from './match.service';
 export * from './player.service';
-export * from './stats.service';
