+++ libs/common/src/utils/index.ts
@@ -1,4 +1,3 @@
 export * from './firestore-db';
 export * from './gemini.helper';
-export * from './stats.util';
 export * from './various';
